# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: LiveTalk
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /rooms/:
    x-swagger-router-controller: room
    post:
      security:
        - Bearer: []
      tags:
        - Room
      description: Create Room
      operationId: createRoom
      parameters:
        - name: room
          in: body
          schema:
            $ref: "#/definitions/RoomContent"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/CommonResult"
    get:
      security:
        - Bearer: []
      tags:
        - Room
      description: list room
      operationId: listRooms
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/CommonListResult"
  /hooks/publish/start:
    x-swagger-router-controller: room

    post:
      consumes:
        - application/x-www-form-urlencoded
      tags:
        - Hook
      parameters:
        - name: name
          in: formData
          type: string
        - name: app
          in: formData
          type: string
        - name: call
          in: formData
          type: string
      description: publish room
      operationId: onPublishStart
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/CommonResult"
  /hooks/publish/stop:
    x-swagger-router-controller: room
    post:
      consumes:
        - application/x-www-form-urlencoded
      tags:
        - Hook
      parameters:
        - name: name
          in: formData
          type: string
        - name: app
          in: formData
          type: string
        - name: call
          in: formData
          type: string
      description: publish room
      operationId: onPublishFinish
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/CommonResult"
  /common/countries/{id}:
    x-swagger-router-controller: common
    get:
      consumes:
        - application/x-www-form-urlencoded
      security:
        - Bearer: []
      tags:
        - Common
      description: get specific country information
      operationId: getCountry
      parameters:
        - name: id
          in: path
          type: string
          required: true
      responses:
        "200":
          description: Get Countries information Success
          schema:
            $ref: "#/definitions/GetCountryResponse"
  /products:
    x-swagger-router-controller: product
    post:
      security:
        - Bearer: []
      tags:
        - Product
      description: create product
      operationId: create
      parameters:
        - name: product
          in: body
          schema:
            $ref: "#/definitions/ProductContent"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ProductCreateResponse"
    get:
      security:
        - Bearer: []
      description: list products
      # used as the method name of the controller
      operationId: list
      parameters:
        - name: name
          in: query
          description: List products....
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/ProductResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  # this show swagger.json for swagger-ui to interpret
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  RoomContent:
    properties:
      name:
        type: string
      description:
        type: string
  CommonResult:
    properties:
      data:
        type: 
          - object
          - 'null'
      status:
        type: integer
      message:
        type: string
  CommonListResult:
    properties:
      data: 
        type: array
        items:
          type: object
      status:
        type: integer
      message:
        type: string
  GetCountryResponse:
    properties:
      name:
        type: string
      location:
        type: string
      latitude:
        type: number
      logitude:
        type: number
  ProductContent:
    properties:
      name:
        type: string
      price:
        type: integer
  ProductCreateResponse:
    properties:
      status:
        type: integer
      message:
        type: string
  ProductResponse:
    properties:
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string